requirements.txt
matplotlib==3.7.2
numpy==1.23.4
openpyxl==3.1.2
PyYAML==6.0.1
scipy==1.12.0
tqdm==4.66.1
